Split xinspect.hpp into interface and implementation
